Refactor HTML generation into separate template class
- Extract all HTML, CSS, and JavaScript into HtmlTemplate class
- Reduce RedirectGenerator complexity from 417 to 155 lines (63% reduction)
- Improve separation of concerns and maintainability
- Keep all functionality intact with cleaner architecture
